+++ src/core/data-snapshot/ChatCommandsData.js
@@ -1,0 +1,9 @@
+const DataSnapshot = require('./DataSnapshot');
+
+class ChatCommandsData extends DataSnapshot {
+    static defaults = {
+        commands: {}
+    };
+}
+
+module.exports = ChatCommandsData;