+++ src/core/data-snapshot/ControlStateData.js
@@ -1,0 +1,10 @@
+const DataSnapshot = require('./DataSnapshot');
+
+class ControlStateData extends DataSnapshot {
+    static defaults = {
+        connections: [],
+        displaySelections: {}
+    };
+}
+
+module.exports = ControlStateData;